change maxsize to an unsigned int
git-svn-id: http://code.sixapart.com/svn/memcached/trunk@29 b0b603af-a30f-0410-a34e-baf09ae79d0b

+++ memcached.h
@@ -19,7 +19,7 @@ struct stats {
 };
 
 struct settings {
-    unsigned long long maxbytes;
+    unsigned int maxbytes;
     int maxconns;
     int port;
     struct in_addr interface;
